% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funKappa.R
\name{funKappa}
\alias{funKappa}
\title{Calculate Cohen's kappa between two vectors}
\usage{
funKappa(vec1, vec2)
}
\arguments{
\item{vec1}{Vector of binary time series (NA not allowed)}

\item{vec2}{Vector of binary time series (equal length as vec1, NA not allowed)}
}
\value{
list with two elements

\code{value} of the Cohen's kappa and

\code{funName} name of the function
}
\description{
Calculate Cohen's kappa between two vectors
}
\examples{
data1<-rep(c(1,0,1,1),25)
        data2<-ifelse(rnorm(100,0,1)<0.7,0,1)
        funKappa(data1,data2)
}
