% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funPhiCC.R
\name{funPhiCC}
\alias{funPhiCC}
\title{Calculate the phi correllation coefficient index between two vectors}
\usage{
funPhiCC(vec1, vec2)
}
\arguments{
\item{vec1}{Vector of binary time series (no missing values)}

\item{vec2}{Vector of binary time series (equal length as vec1, no missing values)}
}
\value{
list with two elements

\code{value} of the phi correlation coefficient and

\code{funName} name of the function
}
\description{
Calculate the phi correllation coefficient index between two vectors
}
\examples{
set.seed(1234)
          data1<-rep(c(1,0,1,1),25)
          data2<-ifelse(rnorm(100,0,1)<0.7,0,1)
          funPhiCC(data1,data2)
}
