% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funPropAgree.R
\name{funPropAgree}
\alias{funPropAgree}
\title{Calculate the proportion of agreement between two vectors}
\usage{
funPropAgree(vec1, vec2)
}
\arguments{
\item{vec1}{Vector of binary time series (no missing values)}

\item{vec2}{Vector of binary time series (equal length as vec1, no missing values)}
}
\value{
list with two elements

\code{value} of the proportion of agreement and

\code{funName} name of the function
}
\description{
Calculate the proportion of agreement between two vectors
}
\examples{
data1 <- rep(c(1,0,1,1),25)
        data2 <- ifelse(rnorm(100,0,1)<0.7,0,1)
        funPropAgree(data1,data2)
}
