% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProb.R
\name{getProb}
\alias{getProb}
\title{Retrieve (conditional) probabilities from binary time series}
\usage{
getProb(ts)
}
\arguments{
\item{ts}{Binary time series vector}
}
\value{
List of three elements

\code{p1} the prevalence p(X(t)=1) and

\code{p1|1} and \code{p1|0} the two auto-conditional probabilities p(X(t)=1|X(t-1)=1) & p(X(t)=1|X(t-1)=0)
}
\description{
Retrieve (conditional) probabilities from binary time series
}
\examples{
getProb(c(1,0,1,0,1,0,1,1,1,1,0,0,0,0,1,0,1,1,0,0,0,0,0,0,0,0))
}
