% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist.conTest.R
\name{hist.conTest}
\alias{hist.conTest}
\title{Plot histogram matrix of the significance test}
\usage{
\method{hist}{conTest}(x, signLev = 0.05, ...)
}
\arguments{
\item{x}{Object of the class conTest}

\item{signLev}{Significance level (default .05)}

\item{...}{Graphical parameters to be passed to hist()}
}
\value{
Histogram matrix with sample distribution and value from observed data for each variable combination
}
\description{
Plot histogram matrix of the significance test
}
\examples{
data(SymptomData)
          test.result <- conTest(SymptomData[,c(2,6,8)],conFun=funClassJacc,
                          typeOfTest='permut',nBlox=10,adCor=TRUE,nReps=100)
          hist(test.result)
}
