% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagthemats.R
\name{lagthemats}
\alias{lagthemats}
\title{Lag a matrix}
\usage{
lagthemats(data, lag)
}
\arguments{
\item{data}{Binary time-points-by-variable matrix}

\item{lag}{Non-negative integer indicating the number of time points the second variable is lagged (default 0)}
}
\value{
\code{laggeddata} Matrix in which all variables are lagged \code{lag} time points
}
\description{
Lag a matrix
}
\examples{
lagthemats(cbind(c(1,0,1,0,1,0,1),c(1,1,1,1,0,0,0)),lag=2)


}
