% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelAD.R
\name{modelAD}
\alias{modelAD}
\title{Generate data with model-based approach accounting for auto-dependence}
\usage{
modelAD(vec)
}
\arguments{
\item{vec}{Time series vector}
}
\value{
Time series vector that is similar to the original one in relative frequency and auto-dependence
}
\description{
This function generates a new time serie that is similar to the original one
in relative frequency and auto-dependence, by drawing samples time point per
time point from a Bernoulli distribution with the different conditional probabilities
as parameter, depending on the state of the previous time point.
}
\examples{
ts=rep(c(1,1,1,1,1,0,0,0,0,0),15)
modelAD(ts)
}
