% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelNO.R
\name{modelNO}
\alias{modelNO}
\title{Generate data with model-based approach ignoring auto-dependence}
\usage{
modelNO(vec)
}
\arguments{
\item{vec}{Time series vector}
}
\value{
Time series vector that is similar to the original one considering relative frequency
}
\description{
This function generates a new time serie that is similar to the original one in relative frequency,
but not in auto-dependence by drawing from a Bernoulli distribution with the relative frequency as parameter.
}
\examples{
ts=rep(c(1,1,1,1,1,0,0,0,0,0),15)
modelNO(ts)

}
