% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutNO.R
\name{permutNO}
\alias{permutNO}
\title{Generate data with permutation-based approach ignoring auto-dependence}
\usage{
permutNO(vec)
}
\arguments{
\item{vec}{Time series vector}
}
\value{
Time series vector with exactly the same relative frequency as the original vector
}
\description{
This function generates a new time serie with exactly the same relative frequency as the original one,
but with a lower auto-dependence by shuffeling all time points.
}
\examples{
ts=rep(c(1,1,1,1,1,0,0,0,0,0),15)
permutNO(ts)
}
