% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.conProf.R
\name{plot.conProf}
\alias{plot.conProf}
\title{Draw contingency profiles}
\usage{
\method{plot}{conProf}(x, ...)
}
\arguments{
\item{x}{Object of a conProf class}

\item{...}{Parameters to be transfered to the plot function}
}
\value{
Contingency profile matrix
}
\description{
Draw contingency profiles
}
\examples{
 IntData <- cbind(rep(rep(c(0,0,1,0,1,0,1,0,0,0),each=5),times=5),
                            rep(rep(c(1,0,0,0), each=10), times=25))
           colnames(IntData) <- c('Var1','Var2')
           CP <- conProf(IntData,maxlag=10,conFun=funClassJacc)
           plot(CP)
}
